+++ src/content.ts
@@ -56,7 +56,7 @@ import { SelectionMeta } from './meta';
 import { kanjiToNumber } from './numbers';
 import { renderPopup, CopyState, PopupOptions } from './popup';
 import { query, QueryResult } from './query';
-import { isEraName } from './years';
+import { isEraName, startsWithEraName } from './years';
 
 declare global {
   interface Window {
@@ -315,7 +315,7 @@ export class RikaiContent {
   detach() {
     window.removeEventListener('mousemove', this.onMouseMove);
     window.removeEventListener('mousedown', this.onMouseDown);
-    window.removeEventListener('keydown', this.onKeyDown);
+    window.removeEventListener('keydown', this.onKeyDown, { capture: true });
     window.removeEventListener('focusin', this.onFocusIn);
 
     this.clearHighlight(null);
@@ -1043,6 +1043,9 @@ export class RikaiContent {
     // Search for non-Japanese text (or a delimiter of some sort even if it
     // is "Japanese" in the sense of being full-width).
     //
+    // * U+FF01~U+FF5E is for full-width alphanumerics (includes some
+    //   punctuation like ＆ and ～ because they appear in the kanji headwords for
+    //   some entries)
     // * U+25CB is 'white circle' often used to represent a blank
     //   (U+3007 is an ideographic zero that is also sometimes used for this
     //   purpose, but this is included in the U+3001~U+30FF range.)
@@ -1080,8 +1083,9 @@ export class RikaiContent {
     //   skip them (although previous rikai-tachi included them) since
     //   they're mostly going to be delimiters
     // * U+FF66~U+FF9F is halfwidth katakana
+    // * U+20000~U+20FFF is CJK Unified Ideographs Extension B (more rare kanji)
     //
-    const nonJapaneseOrDelimiter = /[^\u25cb\u3004-\u3007\u3011-\u30ff\u3220-\u3247\u3280-\u32b0\u32d0-\u32ff\u3300-\u3370\u337b-\u337f\u3400-\u4dbf\u4e00-\u9fff\uf900-\ufaff\uff5e\uff66-\uff9f]/;
+    const nonJapaneseOrDelimiter = /[^\uff01-\uff5e\u25cb\u3004-\u3007\u3011-\u30ff\u3220-\u3247\u3280-\u32b0\u32d0-\u32ff\u3300-\u3370\u337b-\u337f\u3400-\u4dbf\u4e00-\u9fff\uf900-\ufaff\uff5e\uff66-\uff9f\u{20000}-\u{20fff}]/u;
 
     // If we detect a Japanese era, however, we allow a different set of
     // characters.
@@ -1102,7 +1106,7 @@ export class RikaiContent {
 
         // If we hit a delimiter but the existing text is an era name, we should
         // re-find the end of this text node.
-        if (textEnd >= 0 && isEraName(currentText)) {
+        if (textEnd >= 0 && startsWithEraName(currentText)) {
           textDelimiter = nonEraCharacter;
           const endOfEra = nodeText.substring(textEnd).search(textDelimiter);
           textEnd = endOfEra === -1 ? -1 : textEnd + endOfEra;
@@ -1624,7 +1628,6 @@ export class RikaiContent {
       // Add the popup div
       const popup = doc.createElement('div');
       popup.setAttribute('id', 'rikaichamp-window');
-      popup.setAttribute('lang', 'ja');
       popup.classList.add(`-${this._config.popupStyle}`);
       doc.documentElement.append(popup);